* Typecasted time_t references to their appropriate types (%ld being   signed long, %lu being unsigned long).  Yes, there are some %lds,   believe it or not.
git-svn-id: http://svn.ratbox.org/svnroot/trunk@7081 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ contrib/m_clearchan.c
@@ -153,8 +153,11 @@ static void mo_clearchan(struct Client *client_p, struct Client *source_p,
   /* SJOIN the user to give them ops, and lock the channel */
 
   sendto_ll_channel_remote(chptr, client_p, source_p,
-      ":%s SJOIN %lu %s +ntsi :@%s", me.name, (chptr->channelts - 1),
-      chptr->chname, source_p->name);
+      ":%s SJOIN %lu %s +ntsi :@%s",
+      me.name,
+      (unsigned long) (chptr->channelts - 1),
+      chptr->chname,
+      source_p->name);
   sendto_channel_local(ALL_MEMBERS, chptr, ":%s!%s@%s JOIN %s",
                        source_p->name,
                        source_p->username,
