+++ contrib/m_opme.c
@@ -155,20 +155,36 @@ static void mo_opme(struct Client *client_p, struct Client *source_p,
                    parv[1], parv[2], source_p->name, source_p->username, source_p->host);
     }
 
-  sendto_match_cap_servs(chptr, source_p, CAP_UID, ":%s PART %s",
-						 ID(source_p), parv[1]);
-  sendto_match_nocap_servs(chptr, source_p, CAP_UID, ":%s PART %s",
-						   source_p->name, parv[1]);
-  sendto_match_cap_servs(chptr, source_p, CAP_UID, ":%s SJOIN %ld %s + :@%s",
-						 me.name, chptr->channelts, parv[1], /* XXX ID(source_p) */ source_p->name);
-  sendto_match_nocap_servs(chptr, source_p, CAP_UID, ":%s SJOIN %ld %s + :@%s",
-					 me.name, chptr->channelts, parv[1], source_p->name);
-  sendto_channel_local(ALL_MEMBERS, chptr, ":%s MODE %s +o %s",
-					   me.name, parv[1], source_p->name);
+  sendto_match_cap_servs(chptr, source_p, CAP_UID,
+				":%s PART %s",
+				ID(source_p),
+				parv[1]);
+  sendto_match_nocap_servs(chptr, source_p, CAP_UID,
+				":%s PART %s",
+				source_p->name,
+				parv[1]);
+  sendto_match_cap_servs(chptr, source_p, CAP_UID,
+				":%s SJOIN %ld %s + :@%s",
+				me.name,
+				(signed long) chptr->channelts,
+				parv[1],
+				/* XXX ID(source_p) */
+				source_p->name);
+  sendto_match_nocap_servs(chptr, source_p, CAP_UID,
+				":%s SJOIN %ld %s + :@%s",
+				me.name,
+				(signed long) chptr->channelts,
+				parv[1],
+				source_p->name);
+  sendto_channel_local(ALL_MEMBERS, chptr,
+				":%s MODE %s +o %s",
+				me.name,
+				parv[1],
+				source_p->name);
   
 /*
   sendto_channel_remote(chptr, source_p, ":%s MODE %s +o %s",
-						 me.name, parv[1], source_p->name);
+					 me.name, parv[1], source_p->name);
 */
 }
